Broadcom Android Release 7.1 Phase 17.2 (Take 2) code drop

+++ BSEAV/lib/gpu/vc5/platform/nexus/private_nexus.h
@@ -50,7 +50,7 @@ typedef struct
 {
    uint32_t                   magic;
    NEXUS_SurfaceHandle        surface;
-   uint32_t                   physicalOffset;
+   uint64_t                   physicalOffset;
    void                      *cachedPtr;
    int                        fence;
    bool                       poisoned;
